Upgrade to kotlin 1.5

+++ modules/nbt/src/commonMain/kotlin/com/handtruth/mc/nbt/util/NBTBooleanArrayDecoder.kt
@@ -32,7 +32,7 @@ internal class NBTBooleanArrayDecoder(
     }
 
     override fun decodeCharElement(descriptor: SerialDescriptor, index: Int): Char {
-        return decodeByteElement(descriptor, index).toChar()
+        return decodeByteElement(descriptor, index).toInt().toChar()
     }
 
     override fun decodeDoubleElement(descriptor: SerialDescriptor, index: Int): Double {
