+++ modules/nbt/src/commonMain/kotlin/com/handtruth/mc/nbt/util/NBTByteArrayDecoder.kt
@@ -24,7 +24,7 @@ internal class NBTByteArrayDecoder(
     }
 
     override fun decodeCharElement(descriptor: SerialDescriptor, index: Int): Char {
-        return value[index].toChar()
+        return value[index].toInt().toChar()
     }
 
     override fun decodeDoubleElement(descriptor: SerialDescriptor, index: Int): Double {
